scripts: updated deployment scripts
